Implementado parcialmente proyecto preliminar integracion con back
