+++ matecode/backend/parque-industrial/pom.xml
@@ -121,8 +121,7 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration>
                     <source>21</source>
-                    <target>21</target>
-                    <compilerArgs>--enable-preview</compilerArgs>
+					<target>21</target>
                 </configuration>
             </plugin>
         </plugins>
