+++ matecode/backend/parque-industrial/src/main/java/com/parque_industrial/dto/proyecto/CuitRequest.java
@@ -1,0 +1,16 @@
+package com.parque_industrial.dto.proyecto;
+
+public class CuitRequest {
+    private String cuit;
+
+    public CuitRequest() {
+    }
+
+    public String getCuit() {
+        return cuit;
+    }
+
+    public void setCuit(String cuit) {
+        this.cuit = cuit;
+    }
+}
