+++ matecode/frontend/src/features/empresa-dashboard/EnRevision.css
@@ -1,0 +1,71 @@
+.en-revision-container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    min-height: 60vh;
+    padding: 2rem;
+    background-color: #f9fafb;
+}
+
+.card-revision {
+    background: #ffffff;
+    max-width: 500px;
+    width: 100%;
+    padding: 3rem;
+    border-radius: 16px;
+    box-shadow: 0 10px 25px -5px rgba(0, 0, 0, 0.1), 0 8px 10px -6px rgba(0, 0, 0, 0.1);
+    text-align: center;
+    border: 1px solid #e5e7eb;
+}
+
+.icon-box {
+    display: flex;
+    justify-content: center;
+    margin-bottom: 1.5rem;
+    animation: pulse-animation 2s infinite;
+}
+
+.card-revision h2 {
+    color: #111827;
+    font-size: 1.8rem;
+    margin-bottom: 1rem;
+    font-weight: 700;
+}
+
+.card-revision p {
+    color: #4b5563;
+    line-height: 1.6;
+    margin-bottom: 2rem;
+}
+
+.info-box {
+    background-color: #fef3c7;
+    padding: 1rem;
+    border-radius: 8px;
+    border: 1px solid #fde68a;
+}
+
+.info-box p {
+    color: #92400e;
+    font-size: 0.9rem;
+    margin: 0;
+    font-weight: 500;
+}
+
+/* Animación sutil para el icono */
+@keyframes pulse-animation {
+    0% {
+        transform: scale(1);
+        opacity: 1;
+    }
+
+    50% {
+        transform: scale(1.05);
+        opacity: 0.8;
+    }
+
+    100% {
+        transform: scale(1);
+        opacity: 1;
+    }
+}
