+++ matecode/frontend/src/features/empresa-dashboard/ErrorView.tsx
@@ -1,0 +1,5 @@
+export default function ErrorView() {
+    return (
+        <h1>Error</h1>
+    )
+}
