+++ matecode/frontend/src/features/empresa-dashboard/estadosProyecto.tsx
@@ -1,0 +1,11 @@
+export const ESTADOS_PROYECTO = {
+    NINGUNO: 'NINGUNO',
+
+    PRELIMINAR_EN_REVISION: 'PRELIMINAR_EN_REVISION',
+    PRELIMINAR_RECTIFICAR: 'PRELIMINAR_RECTIFICAR',
+    PRELIMINAR_APROBADO: 'PRELIMINAR_APROBADO',
+
+    DEFINITIVO_EN_REVISION: 'DEFINITIVO_EN_REVISION',
+    DEFINITIVO_RECTIFICAR: 'DEFINITIVO_RECTIFICAR',
+    DEFINITIVO_APROBADO: 'DEFINITIVO_APROBADO',
+};
